Monday_indicator/requirements.txt
pandas
ccxt
nest-asyncio
python-dotenv
requests

